backend/requirements.txt
aiofiles==24.1.0
anyio==4.10.0
fastapi==0.116.1
h11==0.16.0
httpx==0.27.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.2
openai==2.1.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.4
pytesseract==0.3.13
python-dotenv==1.1.1
python-magic==0.4.27
python-multipart==0.0.20
PyYAML==6.0.2
reportlab==4.4.3
sniffio==1.3.1
starlette==0.47.3
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
slowapi==0.1.9
limits==5.6.0
celery==5.4.0
redis==5.0.8
prometheus-fastapi-instrumentator==7.0.0
ruff